refactor: remove Mixins::Rule
Remove Minxins::Rule and group search filters in Struts::Filters
