+++ lib/mihari/commands/init.rb
@@ -3,8 +3,6 @@
 module Mihari
   module Commands
     module Initialization
-      include Mixins::Rule
-
       def self.included(thor)
         thor.class_eval do
           desc "rule", "Create a rule file"
@@ -20,6 +18,31 @@ module Mihari
             initialize_rule_yaml filename
 
             Mihari.logger.info "The rule file is initialized as #{filename}."
+          end
+
+          no_commands do
+            #
+            # Returns a template for rule
+            #
+            # @return [String] A template for rule
+            #
+            def rule_template
+              rule = Structs::Rule.from_path_or_id File.expand_path("../templates/rule.yml.erb", __dir__)
+              rule.yaml
+            end
+
+            #
+            # Create (blank) rule file
+            #
+            # @param [String] filename
+            # @param [Dry::Files] files
+            # @param [String] template
+            #
+            # @return [nil]
+            #
+            def initialize_rule_yaml(filename, files = Dry::Files.new, template: rule_template)
+              files.write(filename, template)
+            end
           end
         end
       end
