+++ lib/mihari/web/endpoints/alerts.rb
@@ -43,7 +43,7 @@ module Mihari
           # symbolize hash keys
           filter = filter.to_h.symbolize_keys
 
-          search_filter_with_pagenation = Structs::Alert::SearchFilterWithPagination.new(**filter)
+          search_filter_with_pagenation = Structs::Filters::Alert::SearchFilterWithPagination.new(**filter)
           alerts = Mihari::Alert.search(search_filter_with_pagenation)
           total = Mihari::Alert.count(search_filter_with_pagenation.without_pagination)
