+++ sig/lib/mihari/cli/base.rbs
@@ -3,8 +3,6 @@ module Mihari
     class Base
       include Mihari::Mixins::Hash
 
-      include Mihari::CLI::Mixins::Utils
-
       def self.exit_on_failure?: () -> ::TrueClass
     end
   end
