--- sig/lib/mihari/mixins/rule.rbs
@@ -1,36 +1,0 @@
-module Mihari
-  module Mixins
-    module Rule
-      def load_erb_yaml: (String yaml) -> Hash
-
-      #
-      # Load rule into hash
-      #
-      # @param [String] path_or_id
-      #
-      # @return [Mihari::Structs::Rule::Rule]
-      #
-      def load_rule: (String path_or_id) -> Mihari::Structs::Rule::Rule
-
-      #
-      # Returns a template for rule
-      #
-      # @return [String] A template for rule
-      #
-      def rule_template: () -> String
-
-      def validate_rule!: (Mihari::Structs::Rule::Rule rule) -> nil
-
-      #
-      # Create (blank) rule file
-      #
-      # @param [String] filename
-      # @param [Dry::Files] files
-      # @param [String] template
-      #
-      # @return [nil]
-      #
-      def initialize_rule_yaml: (String filename, ?untyped files, ?template: String template) -> nil
-    end
-  end
-end
