+++ sig/lib/mihari/models/rule.rbs
@@ -10,8 +10,8 @@ module Mihari
 
     def to_h: () -> Hash
 
-    def self.search: (Mihari::Structs::Rule::SearchFilterWithPagination filter) -> Array[Mihari::Rule]
+    def self.search: (Mihari::Structs::Filters::Rule::SearchFilterWithPagination filter) -> Array[Mihari::Rule]
 
-    def self.count: (Mihari::Structs::Rule::SearchFilter filter) -> Integer
+    def self.count: (Mihari::Structs::Filters::Rule::SearchFilter filter) -> Integer
   end
 end
