--- sig/lib/mihari/structs/alert.rbs
@@ -1,27 +1,0 @@
-module Mihari
-  module Structs
-    module Alert
-      class SearchFilter
-        attr_reader artifact_data: String?
-        attr_reader description: String?
-        attr_reader source: String?
-        attr_reader tag_name: String?
-        attr_reader title: String?
-        attr_reader from_at: DateTime?
-        attr_reader to_at: DateTime?
-        attr_reader asn: Integer?
-        attr_reader dns_record: String?
-        attr_reader reverse_dns_name: String?
-
-        def valid_artifact_filters?: () -> bool
-      end
-
-      class SearchFilterWithPagination
-        attr_reader page: Integer
-        attr_reader limit: Integer
-
-        def without_pagination: () -> Mihari::Structs::Alert::SearchFilter
-      end
-    end
-  end
-end
